Move pairing to feature/
