+++ src/colmap/controllers/option_manager.cc
@@ -30,11 +30,11 @@
 #include "colmap/controllers/option_manager.h"
 
 #include "colmap/controllers/feature_extraction.h"
-#include "colmap/controllers/feature_matching.h"
 #include "colmap/controllers/image_reader.h"
 #include "colmap/controllers/incremental_mapper.h"
 #include "colmap/estimators/bundle_adjustment.h"
 #include "colmap/estimators/two_view_geometry.h"
+#include "colmap/feature/pairing.h"
 #include "colmap/feature/sift.h"
 #include "colmap/math/random.h"
 #include "colmap/mvs/fusion.h"
