+++ src/colmap/feature/pairing.h
@@ -29,13 +29,140 @@
 
 #pragma once
 
-#include "colmap/controllers/feature_matching.h"
-#include "colmap/controllers/feature_matching_utils.h"
+#include "colmap/feature/matcher.h"
 #include "colmap/retrieval/visual_index.h"
 #include "colmap/scene/database.h"
+#include "colmap/util/threading.h"
 #include "colmap/util/types.h"
 
 namespace colmap {
+
+struct ExhaustiveMatchingOptions {
+  // Block size, i.e. number of images to simultaneously load into memory.
+  int block_size = 50;
+
+  bool Check() const;
+};
+
+struct VocabTreeMatchingOptions {
+  // Number of images to retrieve for each query image.
+  int num_images = 100;
+
+  // Number of nearest neighbors to retrieve per query feature.
+  int num_nearest_neighbors = 5;
+
+  // Number of nearest-neighbor checks to use in retrieval.
+  int num_checks = 256;
+
+  // How many images to return after spatial verification. Set to 0 to turn off
+  // spatial verification.
+  int num_images_after_verification = 0;
+
+  // The maximum number of features to use for indexing an image. If an
+  // image has more features, only the largest-scale features will be indexed.
+  int max_num_features = -1;
+
+  // Path to the vocabulary tree.
+  std::string vocab_tree_path = "";
+
+  // Optional path to file with specific image names to match.
+  std::string match_list_path = "";
+
+  // Number of threads for indexing and retrieval.
+  int num_threads = -1;
+
+  bool Check() const;
+};
+
+struct SequentialMatchingOptions {
+  // Number of overlapping image pairs.
+  int overlap = 10;
+
+  // Whether to match images against their quadratic neighbors.
+  bool quadratic_overlap = true;
+
+  // Whether to enable vocabulary tree based loop detection.
+  bool loop_detection = false;
+
+  // Loop detection is invoked every `loop_detection_period` images.
+  int loop_detection_period = 10;
+
+  // The number of images to retrieve in loop detection. This number should
+  // be significantly bigger than the sequential matching overlap.
+  int loop_detection_num_images = 50;
+
+  // Number of nearest neighbors to retrieve per query feature.
+  int loop_detection_num_nearest_neighbors = 1;
+
+  // Number of nearest-neighbor checks to use in retrieval.
+  int loop_detection_num_checks = 256;
+
+  // How many images to return after spatial verification. Set to 0 to turn off
+  // spatial verification.
+  int loop_detection_num_images_after_verification = 0;
+
+  // The maximum number of features to use for indexing an image. If an
+  // image has more features, only the largest-scale features will be indexed.
+  int loop_detection_max_num_features = -1;
+
+  // Path to the vocabulary tree.
+  std::string vocab_tree_path = "";
+
+  bool Check() const;
+
+  VocabTreeMatchingOptions VocabTreeOptions() const;
+};
+
+struct SpatialMatchingOptions {
+  // Whether the location priors in the database are GPS coordinates in
+  // the form of longitude and latitude coordinates in degrees.
+  bool is_gps = true;
+
+  // Whether to ignore the Z-component of the location prior.
+  bool ignore_z = true;
+
+  // The maximum number of nearest neighbors to match.
+  int max_num_neighbors = 50;
+
+  // The maximum distance between the query and nearest neighbor. For GPS
+  // coordinates the unit is Euclidean distance in meters.
+  double max_distance = 100;
+
+  // Number of threads for indexing and retrieval.
+  int num_threads = -1;
+
+  bool Check() const;
+};
+
+struct TransitiveMatchingOptions {
+  // The maximum number of image pairs to process in one batch.
+  int batch_size = 1000;
+
+  // The number of transitive closure iterations.
+  int num_iterations = 3;
+
+  bool Check() const;
+};
+
+struct ImagePairsMatchingOptions {
+  // Number of image pairs to match in one batch.
+  int block_size = 1225;
+
+  // Path to the file with the matches.
+  std::string match_list_path = "";
+
+  bool Check() const;
+};
+
+struct FeaturePairsMatchingOptions {
+  // Whether to geometrically verify the given matches.
+  bool verify_matches = true;
+
+  // Path to the file with the matches.
+  std::string match_list_path = "";
+
+  bool Check() const;
+};
 
 class PairGenerator {
  public:
