+++ src/pycolmap/pipeline/match_features.cc
@@ -1,8 +1,8 @@
 #include "colmap/controllers/feature_matching.h"
-#include "colmap/controllers/pairing.h"
 #include "colmap/estimators/two_view_geometry.h"
 #include "colmap/exe/feature.h"
 #include "colmap/exe/sfm.h"
+#include "colmap/feature/pairing.h"
 #include "colmap/feature/sift.h"
 #include "colmap/util/logging.h"
 #include "colmap/util/misc.h"
